requirements.txt
hidapi
pyudev
pillow
